+++ astrodb_utils/__init__.py
@@ -4,12 +4,7 @@ import warnings
 
 from .utils import (  # noqa: F401
     AstroDBError,
-    find_publication,
-    find_source_in_db,
     ingest_instrument,
-    ingest_names,
-    ingest_publication,
-    ingest_source,
     internet_connection,
     load_astrodb,
 )
@@ -25,3 +20,4 @@ handler.setFormatter(LOGFORMAT)
 logger.addHandler(handler)
 
 warnings.filterwarnings("ignore", module="astroquery.simbad")
+
