+++ astrodb_utils/photometry.py
@@ -12,11 +12,11 @@ from astropy.io.votable import parse
 
 from astrodb_utils import (
     AstroDBError,
-    find_publication,
-    find_source_in_db,
     ingest_instrument,
     internet_connection,
 )
+from astrodb_utils.publications import find_publication
+from astrodb_utils.sources import find_source_in_db
 
 logger = logging.getLogger(__name__)
